refactor: refactored Directions folder, removed useless folders and simplified component html
